Refactor: #34 - 카메라와 머지 작업
