+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/Frame/ImageEdit/IEProgressView.swift
@@ -9,12 +9,20 @@ import SwiftUI
 
 struct IEProgressView: View {
     @State var isAnimating = false
-    
     @Binding var isSave:Bool
+    @StateObject var viewModel:IEViewModel
     
     var body: some View {
-        
         VStack{
+            if let compositeImage = viewModel.compositeImage{
+                Image(uiImage: compositeImage)
+                    .resizable()
+                //                    .frame(width: UIScreen.main.bounds.width * 0.8, height: UIScreen.main.bounds.width * 0.8 * (viewModel.frameBGSize.height / viewModel.frameBGSize.width))
+                    .aspectRatio(contentMode: .fit)
+                    .padding()
+            }
+            
+            
             if isSave {
                 Image("check.save")
                     .resizable()
@@ -24,12 +32,11 @@ struct IEProgressView: View {
                     .onAppear{
                         isAnimating = false
                     }
-                Text("저장완료!")
+                Text("갤러리에 저장완료!")
                     .foregroundColor(.pointPink)
                     .fontWeight(.bold)
                 
             } else {
-                
                 Circle()
                     .strokeBorder(style: StrokeStyle(lineWidth: 7, dash: [5]))
                     .frame(width: 50, height: 50)
@@ -50,9 +57,23 @@ struct IEProgressView: View {
             isAnimating = true
             isSave = false
         }
-       
+        
         
     }
 }
 
+struct VisualEffectView: UIViewRepresentable {
+    var effect: UIVisualEffect?
+    var alpha: CGFloat = 0.5 // 기본 알파 값 설정
+    
+    func makeUIView(context: UIViewRepresentableContext<Self>) -> UIVisualEffectView {
+        let visualEffectView = UIVisualEffectView()
+        return visualEffectView
+    }
+    
+    func updateUIView(_ uiView: UIVisualEffectView, context: UIViewRepresentableContext<Self>) {
+        uiView.effect = effect
+        uiView.alpha = alpha // 알파 값 설정
+    }
+}
